requirements.txt
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-API==2.0
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
markdown2==2.3.8
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.22.2
PyMySQL==0.9.3
python-slugify==4.0.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
text-unidecode==1.3
urllib3==1.25.8
uWSGI==2.0.18
Werkzeug==1.0.0
